fix(transform): improve compound path alignment
Match compound subpaths by orientation buckets, rotate closed targets to closest source anchors, and move transform alignment utilities into a dedicated module.

Add regressions for orientation pairing, anchor rotation, and MathTex 1->0 first-frame closeness; update transform metadata assertions to preserved topology semantics.
